Allow the passes icon to be overriden
This introduces a new entry in the `api_keys.yaml` to refer to the icon
file name under the `./images` folder. This entry is optional and will
default to a standard P (for Passgen) icon.
New icons can be mounted inside the container without the need to
include in and rebuild the source. An example swarm mount can be found
in the `docker-compose.yaml`. The README is updated, too.
This closes #1.